+++ FirebaseAI/Tests/Unit/VertexComponentTests.swift
@@ -58,7 +58,7 @@ class VertexComponentTests: XCTestCase {
     XCTAssertEqual(vertex.firebaseInfo.projectID, VertexComponentTests.projectID)
     XCTAssertEqual(vertex.firebaseInfo.apiKey, VertexComponentTests.apiKey)
     XCTAssertEqual(
-      vertex.apiConfig.service, .agentPlatform(
+      vertex.apiConfig.service, .enterprise(
         endpoint: .firebaseProxyProd,
         location: "us-central1"
       )
@@ -77,7 +77,7 @@ class VertexComponentTests: XCTestCase {
     XCTAssertEqual(vertex.firebaseInfo.projectID, VertexComponentTests.projectID)
     XCTAssertEqual(vertex.firebaseInfo.apiKey, VertexComponentTests.apiKey)
     XCTAssertEqual(
-      vertex.apiConfig.service, .agentPlatform(endpoint: .firebaseProxyProd, location: location)
+      vertex.apiConfig.service, .enterprise(endpoint: .firebaseProxyProd, location: location)
     )
     XCTAssertEqual(vertex.apiConfig.service.endpoint, .firebaseProxyProd)
     XCTAssertEqual(vertex.apiConfig.version, .v1beta)
@@ -95,54 +95,54 @@ class VertexComponentTests: XCTestCase {
     XCTAssertEqual(vertex.firebaseInfo.projectID, VertexComponentTests.projectID)
     XCTAssertEqual(vertex.firebaseInfo.apiKey, VertexComponentTests.apiKey)
     XCTAssertEqual(
-      vertex.apiConfig.service, .agentPlatform(endpoint: .firebaseProxyProd, location: location)
+      vertex.apiConfig.service, .enterprise(endpoint: .firebaseProxyProd, location: location)
     )
     XCTAssertEqual(vertex.apiConfig.service.endpoint, .firebaseProxyProd)
     XCTAssertEqual(vertex.apiConfig.version, .v1beta)
   }
 
-  /// Tests that an agentPlatform instance can be created properly using the default Firebase app.
-  func testAgentPlatformInstanceCreation_defaultApp() throws {
-    let vertex = FirebaseAI.firebaseAI(backend: .agentPlatform())
+  /// Tests that an enterprise instance can be created properly using the default Firebase app.
+  func testEnterpriseInstanceCreation_defaultApp() throws {
+    let vertex = FirebaseAI.firebaseAI(backend: .enterprise())
 
     XCTAssertNotNil(vertex)
     XCTAssertEqual(vertex.firebaseInfo.projectID, VertexComponentTests.projectID)
     XCTAssertEqual(vertex.firebaseInfo.apiKey, VertexComponentTests.apiKey)
     XCTAssertEqual(
-      vertex.apiConfig.service, .agentPlatform(endpoint: .firebaseProxyProd, location: "global")
+      vertex.apiConfig.service, .enterprise(endpoint: .firebaseProxyProd, location: "global")
     )
     XCTAssertEqual(vertex.apiConfig.service.endpoint, .firebaseProxyProd)
     XCTAssertEqual(vertex.apiConfig.version, .v1beta)
   }
 
-  /// Tests that an agentPlatform instance can be created properly using the default Firebase app
+  /// Tests that an enterprise instance can be created properly using the default Firebase app
   /// and custom
   /// location.
-  func testAgentPlatformInstanceCreation_defaultApp_customLocation() throws {
-    let vertex = FirebaseAI.firebaseAI(backend: .agentPlatform(location: location))
+  func testEnterpriseInstanceCreation_defaultApp_customLocation() throws {
+    let vertex = FirebaseAI.firebaseAI(backend: .enterprise(location: location))
 
     XCTAssertNotNil(vertex)
     XCTAssertEqual(vertex.firebaseInfo.projectID, VertexComponentTests.projectID)
     XCTAssertEqual(vertex.firebaseInfo.apiKey, VertexComponentTests.apiKey)
     XCTAssertEqual(
-      vertex.apiConfig.service, .agentPlatform(endpoint: .firebaseProxyProd, location: location)
+      vertex.apiConfig.service, .enterprise(endpoint: .firebaseProxyProd, location: location)
     )
     XCTAssertEqual(vertex.apiConfig.service.endpoint, .firebaseProxyProd)
     XCTAssertEqual(vertex.apiConfig.version, .v1beta)
   }
 
-  /// Tests that an agentPlatform instance can be created properly.
-  func testAgentPlatformInstanceCreation_customApp() throws {
+  /// Tests that an enterprise instance can be created properly.
+  func testEnterpriseInstanceCreation_customApp() throws {
     let vertex = FirebaseAI.firebaseAI(
       app: VertexComponentTests.app,
-      backend: .agentPlatform(location: location)
+      backend: .enterprise(location: location)
     )
 
     XCTAssertNotNil(vertex)
     XCTAssertEqual(vertex.firebaseInfo.projectID, VertexComponentTests.projectID)
     XCTAssertEqual(vertex.firebaseInfo.apiKey, VertexComponentTests.apiKey)
     XCTAssertEqual(
-      vertex.apiConfig.service, .agentPlatform(endpoint: .firebaseProxyProd, location: location)
+      vertex.apiConfig.service, .enterprise(endpoint: .firebaseProxyProd, location: location)
     )
     XCTAssertEqual(vertex.apiConfig.service.endpoint, .firebaseProxyProd)
     XCTAssertEqual(vertex.apiConfig.version, .v1beta)
@@ -152,8 +152,8 @@ class VertexComponentTests: XCTestCase {
   func testSameAppAndLocation_instanceReused() throws {
     let app = try XCTUnwrap(VertexComponentTests.app)
 
-    let vertex1 = FirebaseAI.firebaseAI(app: app, backend: .agentPlatform(location: location))
-    let vertex2 = FirebaseAI.firebaseAI(app: app, backend: .agentPlatform(location: location))
+    let vertex1 = FirebaseAI.firebaseAI(app: app, backend: .enterprise(location: location))
+    let vertex2 = FirebaseAI.firebaseAI(app: app, backend: .enterprise(location: location))
 
     // Ensure they're the same instance.
     XCTAssert(vertex1 === vertex2)
@@ -162,11 +162,11 @@ class VertexComponentTests: XCTestCase {
   func testSameAppAndDifferentLocation_newInstanceCreated() throws {
     let vertex1 = FirebaseAI.firebaseAI(
       app: VertexComponentTests.app,
-      backend: .agentPlatform(location: location)
+      backend: .enterprise(location: location)
     )
     let vertex2 = FirebaseAI.firebaseAI(
       app: VertexComponentTests.app,
-      backend: .agentPlatform(location: "differentLocation")
+      backend: .enterprise(location: "differentLocation")
     )
 
     // Ensure they are different instances.
@@ -180,9 +180,9 @@ class VertexComponentTests: XCTestCase {
 
     let vertex1 = FirebaseAI.firebaseAI(
       app: VertexComponentTests.app,
-      backend: .agentPlatform(location: location)
+      backend: .enterprise(location: location)
     )
-    let vertex2 = FirebaseAI.firebaseAI(app: app2, backend: .agentPlatform(location: location))
+    let vertex2 = FirebaseAI.firebaseAI(app: app2, backend: .enterprise(location: location))
 
     XCTAssert(VertexComponentTests.app != app2)
     XCTAssert(vertex1 !== vertex2) // Ensure they are different instances.
@@ -195,11 +195,11 @@ class VertexComponentTests: XCTestCase {
 
     let vertex1 = FirebaseAI.firebaseAI(
       app: VertexComponentTests.app,
-      backend: .agentPlatform(location: location)
+      backend: .enterprise(location: location)
     )
     let vertex2 = FirebaseAI.firebaseAI(
       app: app2,
-      backend: .agentPlatform(location: "differentLocation")
+      backend: .enterprise(location: "differentLocation")
     )
 
     XCTAssert(VertexComponentTests.app != app2)
@@ -210,7 +210,7 @@ class VertexComponentTests: XCTestCase {
     let vertex1 = FirebaseAI.createInstance(
       app: VertexComponentTests.app,
       apiConfig: APIConfig(
-        service: .agentPlatform(endpoint: .firebaseProxyProd, location: location),
+        service: .enterprise(endpoint: .firebaseProxyProd, location: location),
         version: .v1beta
       ),
       useLimitedUseAppCheckTokens: false
@@ -218,7 +218,7 @@ class VertexComponentTests: XCTestCase {
     let vertex2 = FirebaseAI.createInstance(
       app: VertexComponentTests.app,
       apiConfig: APIConfig(
-        service: .agentPlatform(endpoint: .firebaseProxyProd, location: location), version: .v1
+        service: .enterprise(endpoint: .firebaseProxyProd, location: location), version: .v1
       ),
       useLimitedUseAppCheckTokens: false
     )
@@ -241,7 +241,7 @@ class VertexComponentTests: XCTestCase {
       let vertex = FirebaseAI(
         app: app1,
         apiConfig: APIConfig(
-          service: .agentPlatform(endpoint: .firebaseProxyProd, location: "transitory location"),
+          service: .enterprise(endpoint: .firebaseProxyProd, location: "transitory location"),
           version: .v1beta
         ),
         useLimitedUseAppCheckTokens: false
@@ -255,10 +255,10 @@ class VertexComponentTests: XCTestCase {
     XCTAssertNil(weakVertex)
   }
 
-  func testModelResourceName_agentPlatform() throws {
+  func testModelResourceName_enterprise() throws {
     let app = try XCTUnwrap(VertexComponentTests.app)
     let location = "test-location"
-    let vertex = FirebaseAI.firebaseAI(app: app, backend: .agentPlatform(location: location))
+    let vertex = FirebaseAI.firebaseAI(app: app, backend: .enterprise(location: location))
     let model = "test-model-name"
     let projectID = vertex.firebaseInfo.projectID
 
@@ -313,7 +313,7 @@ class VertexComponentTests: XCTestCase {
 
     XCTAssertEqual(generativeModel.modelResourceName, modelResourceName)
     XCTAssertEqual(generativeModel.systemInstruction, expectedSystemInstruction)
-    XCTAssertEqual(generativeModel.apiConfig, FirebaseAI.defaultVertexAIAPIConfig)
+    XCTAssertEqual(generativeModel.apiConfig, FirebaseAI.defaultEnterpriseAPIConfig)
   }
 
   @available(*, deprecated)
@@ -322,7 +322,7 @@ class VertexComponentTests: XCTestCase {
     let vertex = FirebaseAI.firebaseAI(app: app, backend: .vertexAI(location: location))
     let modelResourceName = vertex.modelResourceName(modelName: modelName)
     let expectedAPIConfig = APIConfig(
-      service: .agentPlatform(endpoint: .firebaseProxyProd, location: location), version: .v1beta
+      service: .enterprise(endpoint: .firebaseProxyProd, location: location), version: .v1beta
     )
     let expectedSystemInstruction = ModelContent(role: nil, parts: systemInstruction.parts)
 
@@ -335,12 +335,13 @@ class VertexComponentTests: XCTestCase {
     XCTAssertEqual(generativeModel.apiConfig, expectedAPIConfig)
   }
 
-  func testGenerativeModel_agentPlatform_defaultLocation() async throws {
+  @available(*, deprecated)
+  func testGenerativeModel_deprecatedAgentPlatform_defaultLocation() async throws {
     let app = try XCTUnwrap(VertexComponentTests.app)
     let vertex = FirebaseAI.firebaseAI(app: app, backend: .agentPlatform())
     let modelResourceName = vertex.modelResourceName(modelName: modelName)
     let expectedAPIConfig = APIConfig(
-      service: .agentPlatform(endpoint: .firebaseProxyProd, location: "global"), version: .v1beta
+      service: .enterprise(endpoint: .firebaseProxyProd, location: "global"), version: .v1beta
     )
     let expectedSystemInstruction = ModelContent(role: nil, parts: systemInstruction.parts)
 
@@ -353,12 +354,49 @@ class VertexComponentTests: XCTestCase {
     XCTAssertEqual(generativeModel.apiConfig, expectedAPIConfig)
   }
 
-  func testGenerativeModel_agentPlatform_customLocation() async throws {
+  @available(*, deprecated)
+  func testGenerativeModel_deprecatedAgentPlatform_customLocation() async throws {
     let app = try XCTUnwrap(VertexComponentTests.app)
     let vertex = FirebaseAI.firebaseAI(app: app, backend: .agentPlatform(location: location))
     let modelResourceName = vertex.modelResourceName(modelName: modelName)
     let expectedAPIConfig = APIConfig(
-      service: .agentPlatform(endpoint: .firebaseProxyProd, location: location), version: .v1beta
+      service: .enterprise(endpoint: .firebaseProxyProd, location: location), version: .v1beta
+    )
+    let expectedSystemInstruction = ModelContent(role: nil, parts: systemInstruction.parts)
+
+    let generativeModel = vertex.generativeModel(
+      modelName: modelName, systemInstruction: systemInstruction
+    )
+
+    XCTAssertEqual(generativeModel.modelResourceName, modelResourceName)
+    XCTAssertEqual(generativeModel.systemInstruction, expectedSystemInstruction)
+    XCTAssertEqual(generativeModel.apiConfig, expectedAPIConfig)
+  }
+
+  func testGenerativeModel_enterprise_defaultLocation() async throws {
+    let app = try XCTUnwrap(VertexComponentTests.app)
+    let vertex = FirebaseAI.firebaseAI(app: app, backend: .enterprise())
+    let modelResourceName = vertex.modelResourceName(modelName: modelName)
+    let expectedAPIConfig = APIConfig(
+      service: .enterprise(endpoint: .firebaseProxyProd, location: "global"), version: .v1beta
+    )
+    let expectedSystemInstruction = ModelContent(role: nil, parts: systemInstruction.parts)
+
+    let generativeModel = vertex.generativeModel(
+      modelName: modelName, systemInstruction: systemInstruction
+    )
+
+    XCTAssertEqual(generativeModel.modelResourceName, modelResourceName)
+    XCTAssertEqual(generativeModel.systemInstruction, expectedSystemInstruction)
+    XCTAssertEqual(generativeModel.apiConfig, expectedAPIConfig)
+  }
+
+  func testGenerativeModel_enterprise_customLocation() async throws {
+    let app = try XCTUnwrap(VertexComponentTests.app)
+    let vertex = FirebaseAI.firebaseAI(app: app, backend: .enterprise(location: location))
+    let modelResourceName = vertex.modelResourceName(modelName: modelName)
+    let expectedAPIConfig = APIConfig(
+      service: .enterprise(endpoint: .firebaseProxyProd, location: location), version: .v1beta
     )
     let expectedSystemInstruction = ModelContent(role: nil, parts: systemInstruction.parts)